Moved PollUntil function to utils.go
