+++ examples/resources/nd_backup/resource.tf
@@ -5,4 +5,8 @@ resource "nd_backup" "test_resource_backup_1" {
   destination    = "sftp-server"
   encryption_key = "backupKey123"
   telemetry_data = false
+  timeouts = {
+    create = "90m"
+    read   = "30s"
+  }
 }
